@@ -2,23 +2,23 @@ name: flutter_local_notifications
 description: A cross platform plugin for displaying and scheduling local
   notifications for Flutter applications with the ability to customise for each
   platform.
-version: 4.0.1+2
+version: 5.0.0
 homepage: https://github.com/MaikuB/flutter_local_notifications/tree/master/flutter_local_notifications
 
 dependencies:
   flutter:
     sdk: flutter
-  flutter_local_notifications_platform_interface: ^2.0.0+1
-  platform: ">=2.0.0 <4.0.0"
-  timezone: ^0.6.1
+  flutter_local_notifications_platform_interface: ^3.0.0
+  platform: ^3.0.0
+  timezone: ^0.7.0-nullsafety.0
 
 dev_dependencies:
   flutter_driver:
     sdk: flutter
   flutter_test:
     sdk: flutter
-  mockito: ^4.1.3
-  plugin_platform_interface: ^1.0.3
+  mockito: ^5.0.0-nullsafety.7
+  plugin_platform_interface: ^2.0.0
 
 flutter:
   plugin:
@@ -32,5 +32,5 @@ flutter:
         pluginClass: FlutterLocalNotificationsPlugin
 
 environment:
-  sdk: ">=2.6.0 <3.0.0"
+  sdk: '>=2.12.0-0 <3.0.0'
   flutter: ">=1.12.13+hotfix.5"
