@@ -64,7 +64,7 @@ class FlutterLocalNotificationsPlugin {
   /// type for a result to be returned. For example, when the specified type
   /// argument is of type [AndroidFlutterLocalNotificationsPlugin], this will
   /// only return a result of that type when running on Android.
-  T resolvePlatformSpecificImplementation<
+  T? resolvePlatformSpecificImplementation<
       T extends FlutterLocalNotificationsPlatform>() {
     if (T == FlutterLocalNotificationsPlatform) {
       throw ArgumentError.value(
@@ -76,17 +76,17 @@ class FlutterLocalNotificationsPlugin {
         T == AndroidFlutterLocalNotificationsPlugin &&
         FlutterLocalNotificationsPlatform.instance
             is AndroidFlutterLocalNotificationsPlugin) {
-      return FlutterLocalNotificationsPlatform.instance;
+      return FlutterLocalNotificationsPlatform.instance as T?;
     } else if (_platform.isIOS &&
         T == IOSFlutterLocalNotificationsPlugin &&
         FlutterLocalNotificationsPlatform.instance
             is IOSFlutterLocalNotificationsPlugin) {
-      return FlutterLocalNotificationsPlatform.instance;
+      return FlutterLocalNotificationsPlatform.instance as T?;
     } else if (_platform.isMacOS &&
         T == MacOSFlutterLocalNotificationsPlugin &&
         FlutterLocalNotificationsPlatform.instance
             is MacOSFlutterLocalNotificationsPlugin) {
-      return FlutterLocalNotificationsPlatform.instance;
+      return FlutterLocalNotificationsPlatform.instance as T?;
     }
 
     return null;
@@ -110,24 +110,24 @@ class FlutterLocalNotificationsPlugin {
   /// [IOSInitializationSettings.requestSoundPermission] values to false.
   /// [IOSFlutterLocalNotificationsPlugin.requestPermissions] can then be called
   /// to request permissions when needed.
-  Future<bool> initialize(
+  Future<bool?> initialize(
     InitializationSettings initializationSettings, {
-    SelectNotificationCallback onSelectNotification,
+    SelectNotificationCallback? onSelectNotification,
   }) async {
     if (_platform.isAndroid) {
       return resolvePlatformSpecificImplementation<
               AndroidFlutterLocalNotificationsPlugin>()
-          ?.initialize(initializationSettings?.android,
+          ?.initialize(initializationSettings.android!,
               onSelectNotification: onSelectNotification);
     } else if (_platform.isIOS) {
       return await resolvePlatformSpecificImplementation<
               IOSFlutterLocalNotificationsPlugin>()
-          ?.initialize(initializationSettings?.iOS,
+          ?.initialize(initializationSettings.iOS!,
               onSelectNotification: onSelectNotification);
     } else if (_platform.isMacOS) {
       return await resolvePlatformSpecificImplementation<
               MacOSFlutterLocalNotificationsPlugin>()
-          ?.initialize(initializationSettings?.macOS,
+          ?.initialize(initializationSettings.macOS!,
               onSelectNotification: onSelectNotification);
     }
     return true;
@@ -145,7 +145,8 @@ class FlutterLocalNotificationsPlugin {
   /// Note that this will return null for applications running on macOS
   /// versions older than 10.14. This is because there's currently no mechanism
   /// for plugins to receive information on lifecycle events.
-  Future<NotificationAppLaunchDetails> getNotificationAppLaunchDetails() async {
+  Future<NotificationAppLaunchDetails?>
+      getNotificationAppLaunchDetails() async {
     if (_platform.isAndroid) {
       return await resolvePlatformSpecificImplementation<
               AndroidFlutterLocalNotificationsPlugin>()
@@ -160,7 +161,7 @@ class FlutterLocalNotificationsPlugin {
           ?.getNotificationAppLaunchDetails();
     } else {
       return await FlutterLocalNotificationsPlatform.instance
-              ?.getNotificationAppLaunchDetails() ??
+              .getNotificationAppLaunchDetails() ??
           const NotificationAppLaunchDetails(false, null);
     }
   }
@@ -169,10 +170,10 @@ class FlutterLocalNotificationsPlugin {
   /// the app when a notification is tapped.
   Future<void> show(
     int id,
-    String title,
-    String body,
-    NotificationDetails notificationDetails, {
-    String payload,
+    String? title,
+    String? body,
+    NotificationDetails? notificationDetails, {
+    String? payload,
   }) async {
     if (_platform.isAndroid) {
       await resolvePlatformSpecificImplementation<
@@ -192,7 +193,7 @@ class FlutterLocalNotificationsPlugin {
               notificationDetails: notificationDetails?.macOS,
               payload: payload);
     } else {
-      await FlutterLocalNotificationsPlatform.instance?.show(id, title, body);
+      await FlutterLocalNotificationsPlatform.instance.show(id, title, body);
     }
   }
 
@@ -204,13 +205,13 @@ class FlutterLocalNotificationsPlugin {
   /// The `tag` parameter specifies the Android tag. If it is provided,
   /// then the notification that matches both the id and the tag will
   /// be canceled. `tag` has no effect on other platforms.
-  Future<void> cancel(int id, {String tag}) async {
+  Future<void> cancel(int id, {String? tag}) async {
     if (_platform.isAndroid) {
       await resolvePlatformSpecificImplementation<
               AndroidFlutterLocalNotificationsPlugin>()
           ?.cancel(id, tag: tag);
     } else {
-      await FlutterLocalNotificationsPlatform.instance?.cancel(id);
+      await FlutterLocalNotificationsPlatform.instance.cancel(id);
     }
   }
 
@@ -219,7 +220,7 @@ class FlutterLocalNotificationsPlugin {
   /// This applies to notifications that have been scheduled and those that
   /// have already been presented.
   Future<void> cancelAll() async {
-    await FlutterLocalNotificationsPlatform.instance?.cancelAll();
+    await FlutterLocalNotificationsPlatform.instance.cancelAll();
   }
 
   /// Schedules a notification to be shown at the specified date and time.
@@ -231,23 +232,22 @@ class FlutterLocalNotificationsPlugin {
       'instead.')
   Future<void> schedule(
     int id,
-    String title,
-    String body,
+    String? title,
+    String? body,
     DateTime scheduledDate,
     NotificationDetails notificationDetails, {
-    String payload,
+    String? payload,
     bool androidAllowWhileIdle = false,
   }) async {
     if (_platform.isAndroid) {
       await resolvePlatformSpecificImplementation<
-              AndroidFlutterLocalNotificationsPlugin>()
-          .schedule(
-              id, title, body, scheduledDate, notificationDetails?.android,
+              AndroidFlutterLocalNotificationsPlugin>()!
+          .schedule(id, title, body, scheduledDate, notificationDetails.android,
               payload: payload, androidAllowWhileIdle: androidAllowWhileIdle);
     } else if (_platform.isIOS) {
       await resolvePlatformSpecificImplementation<
               IOSFlutterLocalNotificationsPlugin>()
-          ?.schedule(id, title, body, scheduledDate, notificationDetails?.iOS,
+          ?.schedule(id, title, body, scheduledDate, notificationDetails.iOS,
               payload: payload);
     } else if (_platform.isMacOS) {
       throw UnimplementedError();
@@ -287,22 +287,21 @@ class FlutterLocalNotificationsPlugin {
   /// appear is 2020-10-20 10:00.
   Future<void> zonedSchedule(
     int id,
-    String title,
-    String body,
+    String? title,
+    String? body,
     TZDateTime scheduledDate,
     NotificationDetails notificationDetails, {
-    @required
-        UILocalNotificationDateInterpretation
-            uiLocalNotificationDateInterpretation,
-    @required bool androidAllowWhileIdle,
-    String payload,
-    DateTimeComponents matchDateTimeComponents,
+    required UILocalNotificationDateInterpretation
+        uiLocalNotificationDateInterpretation,
+    required bool androidAllowWhileIdle,
+    String? payload,
+    DateTimeComponents? matchDateTimeComponents,
   }) async {
     if (_platform.isAndroid) {
       await resolvePlatformSpecificImplementation<
-              AndroidFlutterLocalNotificationsPlugin>()
+              AndroidFlutterLocalNotificationsPlugin>()!
           .zonedSchedule(
-              id, title, body, scheduledDate, notificationDetails?.android,
+              id, title, body, scheduledDate, notificationDetails.android,
               payload: payload,
               androidAllowWhileIdle: androidAllowWhileIdle,
               matchDateTimeComponents: matchDateTimeComponents);
@@ -310,7 +309,7 @@ class FlutterLocalNotificationsPlugin {
       await resolvePlatformSpecificImplementation<
               IOSFlutterLocalNotificationsPlugin>()
           ?.zonedSchedule(
-              id, title, body, scheduledDate, notificationDetails?.iOS,
+              id, title, body, scheduledDate, notificationDetails.iOS,
               uiLocalNotificationDateInterpretation:
                   uiLocalNotificationDateInterpretation,
               payload: payload,
@@ -319,7 +318,7 @@ class FlutterLocalNotificationsPlugin {
       await resolvePlatformSpecificImplementation<
               MacOSFlutterLocalNotificationsPlugin>()
           ?.zonedSchedule(
-              id, title, body, scheduledDate, notificationDetails?.macOS,
+              id, title, body, scheduledDate, notificationDetails.macOS,
               payload: payload,
               matchDateTimeComponents: matchDateTimeComponents);
     }
@@ -341,34 +340,33 @@ class FlutterLocalNotificationsPlugin {
   /// repeat.
   Future<void> periodicallyShow(
     int id,
-    String title,
-    String body,
+    String? title,
+    String? body,
     RepeatInterval repeatInterval,
     NotificationDetails notificationDetails, {
-    String payload,
+    String? payload,
     bool androidAllowWhileIdle = false,
   }) async {
     if (_platform.isAndroid) {
       await resolvePlatformSpecificImplementation<
               AndroidFlutterLocalNotificationsPlugin>()
           ?.periodicallyShow(id, title, body, repeatInterval,
-              notificationDetails: notificationDetails?.android,
+              notificationDetails: notificationDetails.android,
               payload: payload,
               androidAllowWhileIdle: androidAllowWhileIdle);
     } else if (_platform.isIOS) {
       await resolvePlatformSpecificImplementation<
               IOSFlutterLocalNotificationsPlugin>()
           ?.periodicallyShow(id, title, body, repeatInterval,
-              notificationDetails: notificationDetails?.iOS, payload: payload);
+              notificationDetails: notificationDetails.iOS, payload: payload);
     } else if (_platform.isMacOS) {
       await resolvePlatformSpecificImplementation<
               MacOSFlutterLocalNotificationsPlugin>()
           ?.periodicallyShow(id, title, body, repeatInterval,
-              notificationDetails: notificationDetails?.macOS,
-              payload: payload);
+              notificationDetails: notificationDetails.macOS, payload: payload);
     } else {
       await FlutterLocalNotificationsPlatform.instance
-          ?.periodicallyShow(id, title, body, repeatInterval);
+          .periodicallyShow(id, title, body, repeatInterval);
     }
   }
 
@@ -380,23 +378,23 @@ class FlutterLocalNotificationsPlugin {
       'matchDateTimeComponents parameter.')
   Future<void> showDailyAtTime(
     int id,
-    String title,
-    String body,
+    String? title,
+    String? body,
     Time notificationTime,
     NotificationDetails notificationDetails, {
-    String payload,
+    String? payload,
   }) async {
     if (_platform.isAndroid) {
       await resolvePlatformSpecificImplementation<
               AndroidFlutterLocalNotificationsPlugin>()
           ?.showDailyAtTime(
-              id, title, body, notificationTime, notificationDetails?.android,
+              id, title, body, notificationTime, notificationDetails.android,
               payload: payload);
     } else if (_platform.isIOS) {
       await resolvePlatformSpecificImplementation<
               IOSFlutterLocalNotificationsPlugin>()
           ?.showDailyAtTime(
-              id, title, body, notificationTime, notificationDetails?.iOS,
+              id, title, body, notificationTime, notificationDetails.iOS,
               payload: payload);
     } else if (_platform.isMacOS) {
       throw UnimplementedError();
@@ -411,24 +409,24 @@ class FlutterLocalNotificationsPlugin {
       'the matchDateTimeComponents parameter.')
   Future<void> showWeeklyAtDayAndTime(
     int id,
-    String title,
-    String body,
+    String? title,
+    String? body,
     Day day,
     Time notificationTime,
     NotificationDetails notificationDetails, {
-    String payload,
+    String? payload,
   }) async {
     if (_platform.isAndroid) {
       await resolvePlatformSpecificImplementation<
               AndroidFlutterLocalNotificationsPlugin>()
           ?.showWeeklyAtDayAndTime(id, title, body, day, notificationTime,
-              notificationDetails?.android,
+              notificationDetails.android,
               payload: payload);
     } else if (_platform.isIOS) {
       await resolvePlatformSpecificImplementation<
               IOSFlutterLocalNotificationsPlugin>()
           ?.showWeeklyAtDayAndTime(
-              id, title, body, day, notificationTime, notificationDetails?.iOS,
+              id, title, body, day, notificationTime, notificationDetails.iOS,
               payload: payload);
     } else if (_platform.isMacOS) {
       throw UnimplementedError();
@@ -437,5 +435,5 @@ class FlutterLocalNotificationsPlugin {
 
   /// Returns a list of notifications pending to be delivered/shown.
   Future<List<PendingNotificationRequest>> pendingNotificationRequests() =>
-      FlutterLocalNotificationsPlatform.instance?.pendingNotificationRequests();
+      FlutterLocalNotificationsPlatform.instance.pendingNotificationRequests();
 }
