@@ -4,9 +4,9 @@ import 'icon.dart';
 class Person {
   /// Constructs an instance of [Person].
   const Person({
-    this.bot,
+    this.bot = false,
     this.icon,
-    this.important,
+    this.important = false,
     this.key,
     this.name,
     this.uri,
@@ -16,17 +16,17 @@ class Person {
   final bool bot;
 
   /// Icon for this person.
-  final AndroidIcon icon;
+  final AndroidIcon? icon;
 
   /// Whether or not this is an important person.
   final bool important;
 
   /// Unique identifier for this person.
-  final String key;
+  final String? key;
 
   /// Name of this person.
-  final String name;
+  final String? name;
 
   /// Uri for this person.
-  final String uri;
+  final String? uri;
 }
