@@ -16,11 +16,11 @@ class InboxStyleInformation extends DefaultStyleInformation {
   }) : super(htmlFormatContent, htmlFormatTitle);
 
   /// Overrides ContentTitle in the big form of the template.
-  final String contentTitle;
+  final String? contentTitle;
 
   /// Set the first line of text after the detail section in the big form of
   /// the template.
-  final String summaryText;
+  final String? summaryText;
 
   /// The lines that form part of the digest section for inbox-style
   /// notifications.
