@@ -4,7 +4,7 @@ class IOSNotificationAttachment {
   const IOSNotificationAttachment(
     this.filePath, {
     this.identifier,
-  }) : assert(filePath != null);
+  });
 
   /// The local file path to the attachment.
   ///
@@ -15,5 +15,5 @@ class IOSNotificationAttachment {
   /// The unique identifier for the attachment.
   ///
   /// When left empty, the iOS APIs will generate a unique identifier
-  final String identifier;
+  final String? identifier;
 }
