@@ -21,7 +21,7 @@ class MacOSNotificationDetails {
   /// to [MacOSInitializationSettings.defaultPresentAlert].
   ///
   /// This property is only applicable to macOS 10.14 or newer.
-  final bool presentAlert;
+  final bool? presentAlert;
 
   /// Play a sound when the notification is triggered while app is in
   /// the foreground.
@@ -30,7 +30,7 @@ class MacOSNotificationDetails {
   /// [MacOSInitializationSettings.defaultPresentSound].
   ///
   /// This property is only applicable to macOS 10.14 or newer.
-  final bool presentSound;
+  final bool? presentSound;
 
   /// Apply the badge value when the notification is triggered while app is in
   /// the foreground.
@@ -39,14 +39,14 @@ class MacOSNotificationDetails {
   /// [MacOSInitializationSettings.defaultPresentBadge].
   ///
   /// This property is only applicable to macOS 10.14 or newer.
-  final bool presentBadge;
+  final bool? presentBadge;
 
   /// Specifies the name of the file to play for the notification.
   ///
   /// Requires setting [presentSound] to true. If [presentSound] is set to true
   /// but [sound] isn't specified then it will use the default notification
   /// sound.
-  final String sound;
+  final String? sound;
 
   /// Specify the number to display as the app icon's badge when the
   /// notification arrives.
@@ -54,19 +54,19 @@ class MacOSNotificationDetails {
   /// Specify the number `0` to remove the current badge, if present. Greater
   /// than `0` to display a badge with that number.
   /// Specify `null` to leave the current badge unchanged.
-  final int badgeNumber;
+  final int? badgeNumber;
 
   /// Specifies the list of attachments included with the notification.
   ///
   /// This property is only applicable to macOS 10.14 or newer.
-  final List<MacOSNotificationAttachment> attachments;
+  final List<MacOSNotificationAttachment>? attachments;
 
   /// Specifies the secondary description.
-  final String subtitle;
+  final String? subtitle;
 
   /// Specifies the thread identifier that can be used to group
   /// notifications together.
   ///
   /// This property is only applicable to macOS 10.14 or newer.
-  final String threadIdentifier;
+  final String? threadIdentifier;
 }
