@@ -12,7 +12,7 @@ import 'package:timezone/timezone.dart' as tz;
 void main() {
   // TODO(maikub): add tests for `periodicallyShow` after https://github.com/dart-lang/sdk/issues/28985 is resolved
   TestWidgetsFlutterBinding.ensureInitialized();
-  FlutterLocalNotificationsPlugin flutterLocalNotificationsPlugin;
+  late FlutterLocalNotificationsPlugin flutterLocalNotificationsPlugin;
 
   group('Android', () {
     const MethodChannel channel =
@@ -26,15 +26,12 @@ void main() {
       channel.setMockMethodCallHandler((methodCall) async {
         log.add(methodCall);
         if (methodCall.method == 'pendingNotificationRequests') {
-          return Future<List<Map<String, Object>>>.value(
-              <Map<String, Object>>[]);
+          return <Map<String, Object?>>[];
         } else if (methodCall.method == 'getNotificationAppLaunchDetails') {
-          return Future<Map<String, Object>>.value(<String, Object>{});
+          return null;
         } else if (methodCall.method == 'getActiveNotifications') {
-          return Future<List<Map<String, Object>>>.value(
-              <Map<String, Object>>[]);
+          return <Map<String, Object?>>[];
         }
-        return Future<void>.value();
       });
     });
 
@@ -65,7 +62,7 @@ void main() {
           1, 'notification title', 'notification body', null);
       expect(
           log.last,
-          isMethodCall('show', arguments: <String, Object>{
+          isMethodCall('show', arguments: <String, Object?>{
             'id': 1,
             'title': 'notification title',
             'body': 'notification body',
@@ -96,7 +93,7 @@ void main() {
             'title': 'notification title',
             'body': 'notification body',
             'payload': '',
-            'platformSpecifics': <String, Object>{
+            'platformSpecifics': <String, Object?>{
               'icon': null,
               'channelId': 'channelId',
               'channelName': 'channelName',
@@ -110,15 +107,15 @@ void main() {
               'enableVibration': true,
               'vibrationPattern': null,
               'groupKey': null,
-              'setAsGroupSummary': null,
+              'setAsGroupSummary': false,
               'groupAlertBehavior': GroupAlertBehavior.all.index,
               'autoCancel': true,
-              'ongoing': null,
+              'ongoing': false,
               'colorAlpha': null,
               'colorRed': null,
               'colorGreen': null,
               'colorBlue': null,
-              'onlyAlertOnce': null,
+              'onlyAlertOnce': false,
               'showWhen': true,
               'when': null,
               'usesChronometer': false,
@@ -175,7 +172,7 @@ void main() {
             'title': 'notification title',
             'body': 'notification body',
             'payload': '',
-            'platformSpecifics': <String, Object>{
+            'platformSpecifics': <String, Object?>{
               'icon': null,
               'channelId': 'channelId',
               'channelName': 'channelName',
@@ -189,15 +186,15 @@ void main() {
               'enableVibration': true,
               'vibrationPattern': null,
               'groupKey': null,
-              'setAsGroupSummary': null,
+              'setAsGroupSummary': false,
               'groupAlertBehavior': GroupAlertBehavior.all.index,
               'autoCancel': true,
-              'ongoing': null,
+              'ongoing': false,
               'colorAlpha': null,
               'colorRed': null,
               'colorGreen': null,
               'colorBlue': null,
-              'onlyAlertOnce': null,
+              'onlyAlertOnce': false,
               'showWhen': true,
               'when': null,
               'usesChronometer': false,
@@ -256,7 +253,7 @@ void main() {
             'title': 'notification title',
             'body': 'notification body',
             'payload': '',
-            'platformSpecifics': <String, Object>{
+            'platformSpecifics': <String, Object?>{
               'icon': null,
               'channelId': 'channelId',
               'channelName': 'channelName',
@@ -270,15 +267,15 @@ void main() {
               'enableVibration': true,
               'vibrationPattern': null,
               'groupKey': null,
-              'setAsGroupSummary': null,
+              'setAsGroupSummary': false,
               'groupAlertBehavior': GroupAlertBehavior.all.index,
               'autoCancel': true,
-              'ongoing': null,
+              'ongoing': false,
               'colorAlpha': null,
               'colorRed': null,
               'colorGreen': null,
               'colorBlue': null,
-              'onlyAlertOnce': null,
+              'onlyAlertOnce': false,
               'showWhen': true,
               'when': timestamp,
               'usesChronometer': false,
@@ -336,7 +333,7 @@ void main() {
             'title': 'notification title',
             'body': 'notification body',
             'payload': '',
-            'platformSpecifics': <String, Object>{
+            'platformSpecifics': <String, Object?>{
               'icon': null,
               'channelId': 'channelId',
               'channelName': 'channelName',
@@ -350,15 +347,15 @@ void main() {
               'enableVibration': true,
               'vibrationPattern': null,
               'groupKey': null,
-              'setAsGroupSummary': null,
+              'setAsGroupSummary': false,
               'groupAlertBehavior': GroupAlertBehavior.all.index,
               'autoCancel': true,
-              'ongoing': null,
+              'ongoing': false,
               'colorAlpha': null,
               'colorRed': null,
               'colorGreen': null,
               'colorBlue': null,
-              'onlyAlertOnce': null,
+              'onlyAlertOnce': false,
               'showWhen': true,
               'when': timestamp,
               'usesChronometer': true,
@@ -419,7 +416,7 @@ void main() {
             'title': 'notification title',
             'body': 'notification body',
             'payload': '',
-            'platformSpecifics': <String, Object>{
+            'platformSpecifics': <String, Object?>{
               'icon': null,
               'channelId': 'channelId',
               'channelName': 'channelName',
@@ -435,15 +432,15 @@ void main() {
               'enableVibration': true,
               'vibrationPattern': null,
               'groupKey': null,
-              'setAsGroupSummary': null,
+              'setAsGroupSummary': false,
               'groupAlertBehavior': GroupAlertBehavior.all.index,
               'autoCancel': true,
-              'ongoing': null,
+              'ongoing': false,
               'colorAlpha': null,
               'colorRed': null,
               'colorGreen': null,
               'colorBlue': null,
-              'onlyAlertOnce': null,
+              'onlyAlertOnce': false,
               'showWhen': true,
               'when': null,
               'usesChronometer': false,
@@ -503,7 +500,7 @@ void main() {
             'title': 'notification title',
             'body': 'notification body',
             'payload': '',
-            'platformSpecifics': <String, Object>{
+            'platformSpecifics': <String, Object?>{
               'icon': null,
               'channelId': 'channelId',
               'channelName': 'channelName',
@@ -519,15 +516,15 @@ void main() {
               'enableVibration': true,
               'vibrationPattern': null,
               'groupKey': null,
-              'setAsGroupSummary': null,
+              'setAsGroupSummary': false,
               'groupAlertBehavior': GroupAlertBehavior.all.index,
               'autoCancel': true,
-              'ongoing': null,
+              'ongoing': false,
               'colorAlpha': null,
               'colorRed': null,
               'colorGreen': null,
               'colorBlue': null,
-              'onlyAlertOnce': null,
+              'onlyAlertOnce': false,
               'showWhen': true,
               'when': null,
               'usesChronometer': false,
@@ -588,7 +585,7 @@ void main() {
             'title': 'notification title',
             'body': 'notification body',
             'payload': '',
-            'platformSpecifics': <String, Object>{
+            'platformSpecifics': <String, Object?>{
               'icon': null,
               'channelId': 'channelId',
               'channelName': 'channelName',
@@ -602,15 +599,15 @@ void main() {
               'enableVibration': true,
               'vibrationPattern': null,
               'groupKey': null,
-              'setAsGroupSummary': null,
+              'setAsGroupSummary': false,
               'groupAlertBehavior': GroupAlertBehavior.all.index,
               'autoCancel': true,
-              'ongoing': null,
+              'ongoing': false,
               'colorAlpha': null,
               'colorRed': null,
               'colorGreen': null,
               'colorBlue': null,
-              'onlyAlertOnce': null,
+              'onlyAlertOnce': false,
               'showWhen': true,
               'when': null,
               'usesChronometer': false,
@@ -673,7 +670,7 @@ void main() {
             'title': 'notification title',
             'body': 'notification body',
             'payload': '',
-            'platformSpecifics': <String, Object>{
+            'platformSpecifics': <String, Object?>{
               'icon': null,
               'channelId': 'channelId',
               'channelName': 'channelName',
@@ -687,15 +684,15 @@ void main() {
               'enableVibration': true,
               'vibrationPattern': null,
               'groupKey': null,
-              'setAsGroupSummary': null,
+              'setAsGroupSummary': false,
               'groupAlertBehavior': GroupAlertBehavior.all.index,
               'autoCancel': true,
-              'ongoing': null,
+              'ongoing': false,
               'colorAlpha': null,
               'colorRed': null,
               'colorGreen': null,
               'colorBlue': null,
-              'onlyAlertOnce': null,
+              'onlyAlertOnce': false,
               'showWhen': true,
               'when': null,
               'usesChronometer': false,
@@ -719,7 +716,7 @@ void main() {
               'shortcutId': null,
               'subText': null,
               'style': AndroidNotificationStyle.bigPicture.index,
-              'styleInformation': <String, Object>{
+              'styleInformation': <String, Object?>{
                 'htmlFormatContent': false,
                 'htmlFormatTitle': false,
                 'bigPicture': 'bigPictureDrawable',
@@ -773,7 +770,7 @@ void main() {
             'title': 'notification title',
             'body': 'notification body',
             'payload': '',
-            'platformSpecifics': <String, Object>{
+            'platformSpecifics': <String, Object?>{
               'icon': null,
               'channelId': 'channelId',
               'channelName': 'channelName',
@@ -787,15 +784,15 @@ void main() {
               'enableVibration': true,
               'vibrationPattern': null,
               'groupKey': null,
-              'setAsGroupSummary': null,
+              'setAsGroupSummary': false,
               'groupAlertBehavior': GroupAlertBehavior.all.index,
               'autoCancel': true,
-              'ongoing': null,
+              'ongoing': false,
               'colorAlpha': null,
               'colorRed': null,
               'colorGreen': null,
               'colorBlue': null,
-              'onlyAlertOnce': null,
+              'onlyAlertOnce': false,
               'showWhen': true,
               'when': null,
               'usesChronometer': false,
@@ -867,7 +864,7 @@ void main() {
             'title': 'notification title',
             'body': 'notification body',
             'payload': '',
-            'platformSpecifics': <String, Object>{
+            'platformSpecifics': <String, Object?>{
               'icon': null,
               'channelId': 'channelId',
               'channelName': 'channelName',
@@ -881,15 +878,15 @@ void main() {
               'enableVibration': true,
               'vibrationPattern': null,
               'groupKey': null,
-              'setAsGroupSummary': null,
+              'setAsGroupSummary': false,
               'groupAlertBehavior': GroupAlertBehavior.all.index,
               'autoCancel': true,
-              'ongoing': null,
+              'ongoing': false,
               'colorAlpha': null,
               'colorRed': null,
               'colorGreen': null,
               'colorBlue': null,
-              'onlyAlertOnce': null,
+              'onlyAlertOnce': false,
               'showWhen': true,
               'when': null,
               'usesChronometer': false,
@@ -913,7 +910,7 @@ void main() {
               'shortcutId': null,
               'subText': null,
               'style': AndroidNotificationStyle.bigPicture.index,
-              'styleInformation': <String, Object>{
+              'styleInformation': <String, Object?>{
                 'htmlFormatContent': false,
                 'htmlFormatTitle': false,
                 'bigPicture': 'bigPictureFilePath',
@@ -967,7 +964,7 @@ void main() {
             'title': 'notification title',
             'body': 'notification body',
             'payload': '',
-            'platformSpecifics': <String, Object>{
+            'platformSpecifics': <String, Object?>{
               'icon': null,
               'channelId': 'channelId',
               'channelName': 'channelName',
@@ -981,15 +978,15 @@ void main() {
               'enableVibration': true,
               'vibrationPattern': null,
               'groupKey': null,
-              'setAsGroupSummary': null,
+              'setAsGroupSummary': false,
               'groupAlertBehavior': GroupAlertBehavior.all.index,
               'autoCancel': true,
-              'ongoing': null,
+              'ongoing': false,
               'colorAlpha': null,
               'colorRed': null,
               'colorGreen': null,
               'colorBlue': null,
-              'onlyAlertOnce': null,
+              'onlyAlertOnce': false,
               'showWhen': true,
               'when': null,
               'usesChronometer': false,
@@ -1059,7 +1056,7 @@ void main() {
             'title': 'notification title',
             'body': 'notification body',
             'payload': '',
-            'platformSpecifics': <String, Object>{
+            'platformSpecifics': <String, Object?>{
               'icon': null,
               'channelId': 'channelId',
               'channelName': 'channelName',
@@ -1073,15 +1070,15 @@ void main() {
               'enableVibration': true,
               'vibrationPattern': null,
               'groupKey': null,
-              'setAsGroupSummary': null,
+              'setAsGroupSummary': false,
               'groupAlertBehavior': GroupAlertBehavior.all.index,
               'autoCancel': true,
-              'ongoing': null,
+              'ongoing': false,
               'colorAlpha': null,
               'colorRed': null,
               'colorGreen': null,
               'colorBlue': null,
-              'onlyAlertOnce': null,
+              'onlyAlertOnce': false,
               'showWhen': true,
               'when': null,
               'usesChronometer': false,
@@ -1105,7 +1102,7 @@ void main() {
               'shortcutId': null,
               'subText': null,
               'style': AndroidNotificationStyle.inbox.index,
-              'styleInformation': <String, Object>{
+              'styleInformation': <String, Object?>{
                 'htmlFormatContent': false,
                 'htmlFormatTitle': false,
                 'lines': <String>['line1'],
@@ -1155,7 +1152,7 @@ void main() {
             'title': 'notification title',
             'body': 'notification body',
             'payload': '',
-            'platformSpecifics': <String, Object>{
+            'platformSpecifics': <String, Object?>{
               'icon': null,
               'channelId': 'channelId',
               'channelName': 'channelName',
@@ -1169,15 +1166,15 @@ void main() {
               'enableVibration': true,
               'vibrationPattern': null,
               'groupKey': null,
-              'setAsGroupSummary': null,
+              'setAsGroupSummary': false,
               'groupAlertBehavior': GroupAlertBehavior.all.index,
               'autoCancel': true,
-              'ongoing': null,
+              'ongoing': false,
               'colorAlpha': null,
               'colorRed': null,
               'colorGreen': null,
               'colorBlue': null,
-              'onlyAlertOnce': null,
+              'onlyAlertOnce': false,
               'showWhen': true,
               'when': null,
               'usesChronometer': false,
@@ -1242,7 +1239,7 @@ void main() {
             'title': 'notification title',
             'body': 'notification body',
             'payload': '',
-            'platformSpecifics': <String, Object>{
+            'platformSpecifics': <String, Object?>{
               'icon': null,
               'channelId': 'channelId',
               'channelName': 'channelName',
@@ -1256,15 +1253,15 @@ void main() {
               'enableVibration': true,
               'vibrationPattern': null,
               'groupKey': null,
-              'setAsGroupSummary': null,
+              'setAsGroupSummary': false,
               'groupAlertBehavior': GroupAlertBehavior.all.index,
               'autoCancel': true,
-              'ongoing': null,
+              'ongoing': false,
               'colorAlpha': null,
               'colorRed': null,
               'colorGreen': null,
               'colorBlue': null,
-              'onlyAlertOnce': null,
+              'onlyAlertOnce': false,
               'showWhen': true,
               'when': null,
               'usesChronometer': false,
@@ -1326,7 +1323,7 @@ void main() {
             'title': 'notification title',
             'body': 'notification body',
             'payload': '',
-            'platformSpecifics': <String, Object>{
+            'platformSpecifics': <String, Object?>{
               'icon': null,
               'channelId': 'channelId',
               'channelName': 'channelName',
@@ -1340,15 +1337,15 @@ void main() {
               'enableVibration': true,
               'vibrationPattern': null,
               'groupKey': null,
-              'setAsGroupSummary': null,
+              'setAsGroupSummary': false,
               'groupAlertBehavior': GroupAlertBehavior.all.index,
               'autoCancel': true,
-              'ongoing': null,
+              'ongoing': false,
               'colorAlpha': null,
               'colorRed': null,
               'colorGreen': null,
               'colorBlue': null,
-              'onlyAlertOnce': null,
+              'onlyAlertOnce': false,
               'showWhen': true,
               'when': null,
               'usesChronometer': false,
@@ -1417,7 +1414,7 @@ void main() {
             'title': 'notification title',
             'body': 'notification body',
             'payload': '',
-            'platformSpecifics': <String, Object>{
+            'platformSpecifics': <String, Object?>{
               'icon': null,
               'channelId': 'channelId',
               'channelName': 'channelName',
@@ -1431,15 +1428,15 @@ void main() {
               'enableVibration': true,
               'vibrationPattern': null,
               'groupKey': null,
-              'setAsGroupSummary': null,
+              'setAsGroupSummary': false,
               'groupAlertBehavior': GroupAlertBehavior.all.index,
               'autoCancel': true,
-              'ongoing': null,
+              'ongoing': false,
               'colorAlpha': null,
               'colorRed': null,
               'colorGreen': null,
               'colorBlue': null,
-              'onlyAlertOnce': null,
+              'onlyAlertOnce': false,
               'showWhen': true,
               'when': null,
               'usesChronometer': false,
@@ -1463,20 +1460,20 @@ void main() {
               'shortcutId': null,
               'subText': null,
               'style': AndroidNotificationStyle.messaging.index,
-              'styleInformation': <String, Object>{
+              'styleInformation': <String, Object?>{
                 'htmlFormatContent': false,
                 'htmlFormatTitle': false,
-                'person': <String, Object>{
-                  'bot': null,
-                  'important': null,
+                'person': <String, Object?>{
+                  'bot': false,
+                  'important': false,
                   'key': null,
                   'name': 'name',
                   'uri': null,
                 },
                 'conversationTitle': null,
                 'groupConversation': null,
-                'messages': <Map<String, Object>>[
-                  <String, Object>{
+                'messages': <Map<String, Object?>>[
+                  <String, Object?>{
                     'text': 'message 1',
                     'timestamp': messageDateTime.millisecondsSinceEpoch,
                     'person': null,
@@ -1537,7 +1534,7 @@ void main() {
             'title': 'notification title',
             'body': 'notification body',
             'payload': '',
-            'platformSpecifics': <String, Object>{
+            'platformSpecifics': <String, Object?>{
               'icon': null,
               'channelId': 'channelId',
               'channelName': 'channelName',
@@ -1551,15 +1548,15 @@ void main() {
               'enableVibration': true,
               'vibrationPattern': null,
               'groupKey': null,
-              'setAsGroupSummary': null,
+              'setAsGroupSummary': false,
               'groupAlertBehavior': GroupAlertBehavior.all.index,
               'autoCancel': true,
-              'ongoing': null,
+              'ongoing': false,
               'colorAlpha': null,
               'colorRed': null,
               'colorGreen': null,
               'colorBlue': null,
-              'onlyAlertOnce': null,
+              'onlyAlertOnce': false,
               'showWhen': true,
               'when': null,
               'usesChronometer': false,
@@ -1597,8 +1594,8 @@ void main() {
                 },
                 'conversationTitle': 'conversationTitle',
                 'groupConversation': true,
-                'messages': <Map<String, Object>>[
-                  <String, Object>{
+                'messages': <Map<String, Object?>>[
+                  <String, Object?>{
                     'text': 'message 1',
                     'timestamp': messageDateTime.millisecondsSinceEpoch,
                     'person': null,
@@ -1645,7 +1642,7 @@ void main() {
               'payload': '',
               'timeZoneName': 'Australia/Sydney',
               'scheduledDateTime': _convertDateToISO8601String(scheduledDate),
-              'platformSpecifics': <String, Object>{
+              'platformSpecifics': <String, Object?>{
                 'allowWhileIdle': true,
                 'icon': null,
                 'channelId': 'channelId',
@@ -1660,15 +1657,15 @@ void main() {
                 'enableVibration': true,
                 'vibrationPattern': null,
                 'groupKey': null,
-                'setAsGroupSummary': null,
+                'setAsGroupSummary': false,
                 'groupAlertBehavior': GroupAlertBehavior.all.index,
                 'autoCancel': true,
-                'ongoing': null,
+                'ongoing': false,
                 'colorAlpha': null,
                 'colorRed': null,
                 'colorGreen': null,
                 'colorBlue': null,
-                'onlyAlertOnce': null,
+                'onlyAlertOnce': false,
                 'showWhen': true,
                 'when': null,
                 'usesChronometer': false,
@@ -1736,7 +1733,7 @@ void main() {
               'timeZoneName': 'Australia/Sydney',
               'scheduledDateTime': _convertDateToISO8601String(scheduledDate),
               'matchDateTimeComponents': DateTimeComponents.time.index,
-              'platformSpecifics': <String, Object>{
+              'platformSpecifics': <String, Object?>{
                 'allowWhileIdle': true,
                 'icon': null,
                 'channelId': 'channelId',
@@ -1751,15 +1748,15 @@ void main() {
                 'enableVibration': true,
                 'vibrationPattern': null,
                 'groupKey': null,
-                'setAsGroupSummary': null,
+                'setAsGroupSummary': false,
                 'groupAlertBehavior': GroupAlertBehavior.all.index,
                 'autoCancel': true,
-                'ongoing': null,
+                'ongoing': false,
                 'colorAlpha': null,
                 'colorRed': null,
                 'colorGreen': null,
                 'colorBlue': null,
-                'onlyAlertOnce': null,
+                'onlyAlertOnce': false,
                 'showWhen': true,
                 'when': null,
                 'usesChronometer': false,
@@ -1828,7 +1825,7 @@ void main() {
               'scheduledDateTime': _convertDateToISO8601String(scheduledDate),
               'matchDateTimeComponents':
                   DateTimeComponents.dayOfWeekAndTime.index,
-              'platformSpecifics': <String, Object>{
+              'platformSpecifics': <String, Object?>{
                 'allowWhileIdle': true,
                 'icon': null,
                 'channelId': 'channelId',
@@ -1843,15 +1840,15 @@ void main() {
                 'enableVibration': true,
                 'vibrationPattern': null,
                 'groupKey': null,
-                'setAsGroupSummary': null,
+                'setAsGroupSummary': false,
                 'groupAlertBehavior': GroupAlertBehavior.all.index,
                 'autoCancel': true,
-                'ongoing': null,
+                'ongoing': false,
                 'colorAlpha': null,
                 'colorRed': null,
                 'colorGreen': null,
                 'colorBlue': null,
-                'onlyAlertOnce': null,
+                'onlyAlertOnce': false,
                 'showWhen': true,
                 'when': null,
                 'usesChronometer': false,
@@ -1889,12 +1886,12 @@ void main() {
       test('without description', () async {
         await flutterLocalNotificationsPlugin
             .resolvePlatformSpecificImplementation<
-                AndroidFlutterLocalNotificationsPlugin>()
+                AndroidFlutterLocalNotificationsPlugin>()!
             .createNotificationChannelGroup(
                 const AndroidNotificationChannelGroup('groupId', 'groupName'));
         expect(log, <Matcher>[
           isMethodCall('createNotificationChannelGroup',
-              arguments: <String, Object>{
+              arguments: <String, Object?>{
                 'id': 'groupId',
                 'name': 'groupName',
                 'description': null,
@@ -1904,7 +1901,7 @@ void main() {
       test('with description', () async {
         await flutterLocalNotificationsPlugin
             .resolvePlatformSpecificImplementation<
-                AndroidFlutterLocalNotificationsPlugin>()
+                AndroidFlutterLocalNotificationsPlugin>()!
             .createNotificationChannelGroup(
                 const AndroidNotificationChannelGroup('groupId', 'groupName',
                     description: 'groupDescription'));
@@ -1922,11 +1919,11 @@ void main() {
     test('createNotificationChannel with default settings', () async {
       await flutterLocalNotificationsPlugin
           .resolvePlatformSpecificImplementation<
-              AndroidFlutterLocalNotificationsPlugin>()
+              AndroidFlutterLocalNotificationsPlugin>()!
           .createNotificationChannel(const AndroidNotificationChannel(
               'channelId', 'channelName', 'channelDescription'));
       expect(log, <Matcher>[
-        isMethodCall('createNotificationChannel', arguments: <String, Object>{
+        isMethodCall('createNotificationChannel', arguments: <String, Object?>{
           'id': 'channelId',
           'name': 'channelName',
           'description': 'channelDescription',
@@ -1942,7 +1939,7 @@ void main() {
           'ledColorGreen': null,
           'ledColorBlue': null,
           'channelAction':
-              AndroidNotificationChannelAction.createIfNotExists?.index,
+              AndroidNotificationChannelAction.createIfNotExists.index,
         })
       ]);
     });
@@ -1950,7 +1947,7 @@ void main() {
     test('createNotificationChannel with non-default settings', () async {
       await flutterLocalNotificationsPlugin
           .resolvePlatformSpecificImplementation<
-              AndroidFlutterLocalNotificationsPlugin>()
+              AndroidFlutterLocalNotificationsPlugin>()!
           .createNotificationChannel(const AndroidNotificationChannel(
             'channelId',
             'channelName',
@@ -1964,7 +1961,7 @@ void main() {
             ledColor: Color.fromARGB(255, 255, 0, 0),
           ));
       expect(log, <Matcher>[
-        isMethodCall('createNotificationChannel', arguments: <String, Object>{
+        isMethodCall('createNotificationChannel', arguments: <String, Object?>{
           'id': 'channelId',
           'name': 'channelName',
           'description': 'channelDescription',
@@ -1980,7 +1977,7 @@ void main() {
           'ledColorGreen': 0,
           'ledColorBlue': 0,
           'channelAction':
-              AndroidNotificationChannelAction.createIfNotExists?.index,
+              AndroidNotificationChannelAction.createIfNotExists.index,
         })
       ]);
     });
@@ -1988,7 +1985,7 @@ void main() {
     test('deleteNotificationChannel', () async {
       await flutterLocalNotificationsPlugin
           .resolvePlatformSpecificImplementation<
-              AndroidFlutterLocalNotificationsPlugin>()
+              AndroidFlutterLocalNotificationsPlugin>()!
           .deleteNotificationChannel('channelId');
       expect(log, <Matcher>[
         isMethodCall('deleteNotificationChannel', arguments: 'channelId')
@@ -1998,7 +1995,7 @@ void main() {
     test('getActiveNotifications', () async {
       await flutterLocalNotificationsPlugin
           .resolvePlatformSpecificImplementation<
-              AndroidFlutterLocalNotificationsPlugin>()
+              AndroidFlutterLocalNotificationsPlugin>()!
           .getActiveNotifications();
       expect(log,
           <Matcher>[isMethodCall('getActiveNotifications', arguments: null)]);
@@ -2007,7 +2004,7 @@ void main() {
     test('cancel', () async {
       await flutterLocalNotificationsPlugin.cancel(1);
       expect(log, <Matcher>[
-        isMethodCall('cancel', arguments: <String, Object>{
+        isMethodCall('cancel', arguments: <String, Object?>{
           'id': 1,
           'tag': null,
         })
@@ -2040,7 +2037,7 @@ void main() {
       debugDefaultTargetPlatformOverride = TargetPlatform.android;
       await flutterLocalNotificationsPlugin
           .resolvePlatformSpecificImplementation<
-              AndroidFlutterLocalNotificationsPlugin>()
+              AndroidFlutterLocalNotificationsPlugin>()!
           .getActiveNotifications();
       expect(log,
           <Matcher>[isMethodCall('getActiveNotifications', arguments: null)]);
@@ -2063,15 +2060,13 @@ void main() {
       flutterLocalNotificationsPlugin = FlutterLocalNotificationsPlugin.private(
           FakePlatform(operatingSystem: 'ios'));
       // ignore: always_specify_types
-      channel.setMockMethodCallHandler((methodCall) {
+      channel.setMockMethodCallHandler((methodCall) async {
         log.add(methodCall);
         if (methodCall.method == 'pendingNotificationRequests') {
-          return Future<List<Map<String, Object>>>.value(
-              <Map<String, Object>>[]);
+          return <Map<String, Object>?>[];
         } else if (methodCall.method == 'getNotificationAppLaunchDetails') {
-          return Future<Map<String, Object>>.value(<String, Object>{});
+          return null;
         }
-        return Future<void>.value();
       });
     });
 
@@ -2131,7 +2126,7 @@ void main() {
           1, 'notification title', 'notification body', null);
       expect(
           log.last,
-          isMethodCall('show', arguments: <String, Object>{
+          isMethodCall('show', arguments: <String, Object?>{
             'id': 1,
             'title': 'notification title',
             'body': 'notification body',
@@ -2169,7 +2164,7 @@ void main() {
             'title': 'notification title',
             'body': 'notification body',
             'payload': '',
-            'platformSpecifics': <String, Object>{
+            'platformSpecifics': <String, Object?>{
               'presentAlert': true,
               'presentBadge': true,
               'presentSound': true,
@@ -2232,7 +2227,7 @@ void main() {
                   UILocalNotificationDateInterpretation.absoluteTime.index,
               'scheduledDateTime': _convertDateToISO8601String(scheduledDate),
               'timeZoneName': 'Australia/Sydney',
-              'platformSpecifics': <String, Object>{
+              'platformSpecifics': <String, Object?>{
                 'presentAlert': true,
                 'presentBadge': true,
                 'presentSound': true,
@@ -2296,7 +2291,7 @@ void main() {
               'scheduledDateTime': _convertDateToISO8601String(scheduledDate),
               'timeZoneName': 'Australia/Sydney',
               'matchDateTimeComponents': DateTimeComponents.time.index,
-              'platformSpecifics': <String, Object>{
+              'platformSpecifics': <String, Object?>{
                 'presentAlert': true,
                 'presentBadge': true,
                 'presentSound': true,
@@ -2361,7 +2356,7 @@ void main() {
               'timeZoneName': 'Australia/Sydney',
               'matchDateTimeComponents':
                   DateTimeComponents.dayOfWeekAndTime.index,
-              'platformSpecifics': <String, Object>{
+              'platformSpecifics': <String, Object?>{
                 'presentAlert': true,
                 'presentBadge': true,
                 'presentSound': true,
@@ -2382,20 +2377,20 @@ void main() {
     test('requestPermissions with default settings', () async {
       await flutterLocalNotificationsPlugin
           .resolvePlatformSpecificImplementation<
-              IOSFlutterLocalNotificationsPlugin>()
+              IOSFlutterLocalNotificationsPlugin>()!
           .requestPermissions();
       expect(log, <Matcher>[
-        isMethodCall('requestPermissions', arguments: <String, Object>{
-          'sound': null,
-          'badge': null,
-          'alert': null,
+        isMethodCall('requestPermissions', arguments: <String, Object?>{
+          'sound': false,
+          'badge': false,
+          'alert': false,
         })
       ]);
     });
     test('requestPermissions with all settings requested', () async {
       await flutterLocalNotificationsPlugin
           .resolvePlatformSpecificImplementation<
-              IOSFlutterLocalNotificationsPlugin>()
+              IOSFlutterLocalNotificationsPlugin>()!
           .requestPermissions(sound: true, badge: true, alert: true);
       expect(log, <Matcher>[
         isMethodCall('requestPermissions', arguments: <String, Object>{
@@ -2439,15 +2434,13 @@ void main() {
       flutterLocalNotificationsPlugin = FlutterLocalNotificationsPlugin.private(
           FakePlatform(operatingSystem: 'macos'));
       // ignore: always_specify_types
-      channel.setMockMethodCallHandler((methodCall) {
+      channel.setMockMethodCallHandler((methodCall) async {
         log.add(methodCall);
         if (methodCall.method == 'pendingNotificationRequests') {
-          return Future<List<Map<String, Object>>>.value(
-              <Map<String, Object>>[]);
+          return <Map<String, Object>?>[];
         } else if (methodCall.method == 'getNotificationAppLaunchDetails') {
-          return Future<Map<String, Object>>.value(<String, Object>{});
+          return null;
         }
-        return Future<void>.value();
       });
     });
 
@@ -2507,7 +2500,7 @@ void main() {
           1, 'notification title', 'notification body', null);
       expect(
           log.last,
-          isMethodCall('show', arguments: <String, Object>{
+          isMethodCall('show', arguments: <String, Object?>{
             'id': 1,
             'title': 'notification title',
             'body': 'notification body',
@@ -2607,7 +2600,7 @@ void main() {
               'payload': '',
               'scheduledDateTime': _convertDateToISO8601String(scheduledDate),
               'timeZoneName': 'Australia/Sydney',
-              'platformSpecifics': <String, Object>{
+              'platformSpecifics': <String, Object?>{
                 'subtitle': null,
                 'presentAlert': true,
                 'presentBadge': true,
@@ -2669,7 +2662,7 @@ void main() {
               'scheduledDateTime': _convertDateToISO8601String(scheduledDate),
               'timeZoneName': 'Australia/Sydney',
               'matchDateTimeComponents': DateTimeComponents.time.index,
-              'platformSpecifics': <String, Object>{
+              'platformSpecifics': <String, Object?>{
                 'subtitle': null,
                 'presentAlert': true,
                 'presentBadge': true,
@@ -2732,7 +2725,7 @@ void main() {
               'timeZoneName': 'Australia/Sydney',
               'matchDateTimeComponents':
                   DateTimeComponents.dayOfWeekAndTime.index,
-              'platformSpecifics': <String, Object>{
+              'platformSpecifics': <String, Object?>{
                 'subtitle': null,
                 'presentAlert': true,
                 'presentBadge': true,
@@ -2754,10 +2747,10 @@ void main() {
     test('requestPermissions with default settings', () async {
       await flutterLocalNotificationsPlugin
           .resolvePlatformSpecificImplementation<
-              MacOSFlutterLocalNotificationsPlugin>()
+              MacOSFlutterLocalNotificationsPlugin>()!
           .requestPermissions();
       expect(log, <Matcher>[
-        isMethodCall('requestPermissions', arguments: <String, Object>{
+        isMethodCall('requestPermissions', arguments: <String, Object?>{
           'sound': null,
           'badge': null,
           'alert': null,
@@ -2767,7 +2760,7 @@ void main() {
     test('requestPermissions with all settings requested', () async {
       await flutterLocalNotificationsPlugin
           .resolvePlatformSpecificImplementation<
-              MacOSFlutterLocalNotificationsPlugin>()
+              MacOSFlutterLocalNotificationsPlugin>()!
           .requestPermissions(sound: true, badge: true, alert: true);
       expect(log, <Matcher>[
         isMethodCall('requestPermissions', arguments: <String, Object>{
