@@ -21,7 +21,7 @@ extension AndroidInitializationSettingsMapper on AndroidInitializationSettings {
 }
 
 extension MessageMapper on Message {
-  Map<String, Object> toMap() => <String, Object>{
+  Map<String, Object?> toMap() => <String, Object?>{
         'text': text,
         'timestamp': timestamp.millisecondsSinceEpoch,
         'person': person?.toMap(),
@@ -32,7 +32,7 @@ extension MessageMapper on Message {
 
 extension AndroidNotificationChannelGroupMapper
     on AndroidNotificationChannelGroup {
-  Map<String, Object> toMap() => <String, Object>{
+  Map<String, Object?> toMap() => <String, Object?>{
         'id': id,
         'name': name,
         'description': description,
@@ -40,7 +40,7 @@ extension AndroidNotificationChannelGroupMapper
 }
 
 extension AndroidNotificationChannelMapper on AndroidNotificationChannel {
-  Map<String, Object> toMap() => <String, Object>{
+  Map<String, Object?> toMap() => <String, Object?>{
         'id': id,
         'name': name,
         'description': description,
@@ -56,12 +56,12 @@ extension AndroidNotificationChannelMapper on AndroidNotificationChannel {
         'ledColorGreen': ledColor?.green,
         'ledColorBlue': ledColor?.blue,
         'channelAction':
-            AndroidNotificationChannelAction.createIfNotExists?.index,
+            AndroidNotificationChannelAction.createIfNotExists.index,
       }..addAll(_convertNotificationSoundToMap(sound));
 }
 
 Map<String, Object> _convertNotificationSoundToMap(
-    AndroidNotificationSound sound) {
+    AndroidNotificationSound? sound) {
   if (sound is RawResourceAndroidNotificationSound) {
     return <String, Object>{
       'sound': sound.sound,
@@ -78,7 +78,7 @@ Map<String, Object> _convertNotificationSoundToMap(
 }
 
 extension PersonMapper on Person {
-  Map<String, Object> toMap() => <String, Object>{
+  Map<String, Object?> toMap() => <String, Object?>{
         'bot': bot,
         'important': important,
         'key': key,
@@ -89,22 +89,22 @@ extension PersonMapper on Person {
   Map<String, Object> _convertIconToMap() {
     if (icon is DrawableResourceAndroidIcon) {
       return <String, Object>{
-        'icon': icon.icon,
+        'icon': icon!.icon,
         'iconSource': AndroidIconSource.drawableResource.index,
       };
     } else if (icon is BitmapFilePathAndroidIcon) {
       return <String, Object>{
-        'icon': icon.icon,
+        'icon': icon!.icon,
         'iconSource': AndroidIconSource.bitmapFilePath.index,
       };
     } else if (icon is ContentUriAndroidIcon) {
       return <String, Object>{
-        'icon': icon.icon,
+        'icon': icon!.icon,
         'iconSource': AndroidIconSource.contentUri.index,
       };
     } else if (icon is FlutterBitmapAssetAndroidIcon) {
       return <String, Object>{
-        'icon': icon.icon,
+        'icon': icon!.icon,
         'iconSource': AndroidIconSource.flutterBitmapAsset.index,
       };
     } else {
@@ -114,21 +114,21 @@ extension PersonMapper on Person {
 }
 
 extension DefaultStyleInformationMapper on DefaultStyleInformation {
-  Map<String, Object> toMap() => _convertDefaultStyleInformationToMap(this);
+  Map<String, Object?> toMap() => _convertDefaultStyleInformationToMap(this);
 }
 
-Map<String, Object> _convertDefaultStyleInformationToMap(
+Map<String, Object?> _convertDefaultStyleInformationToMap(
         DefaultStyleInformation styleInformation) =>
-    <String, Object>{
+    <String, Object?>{
       'htmlFormatContent': styleInformation.htmlFormatContent,
       'htmlFormatTitle': styleInformation.htmlFormatTitle
     };
 
 extension BigPictureStyleInformationMapper on BigPictureStyleInformation {
-  Map<String, Object> toMap() => _convertDefaultStyleInformationToMap(this)
+  Map<String, Object?> toMap() => _convertDefaultStyleInformationToMap(this)
     ..addAll(_convertBigPictureToMap())
     ..addAll(_convertLargeIconToMap())
-    ..addAll(<String, Object>{
+    ..addAll(<String, Object?>{
       'contentTitle': contentTitle,
       'summaryText': summaryText,
       'htmlFormatContentTitle': htmlFormatContentTitle,
@@ -155,12 +155,12 @@ extension BigPictureStyleInformationMapper on BigPictureStyleInformation {
   Map<String, Object> _convertLargeIconToMap() {
     if (largeIcon is DrawableResourceAndroidBitmap) {
       return <String, Object>{
-        'largeIcon': largeIcon.bitmap,
+        'largeIcon': largeIcon!.bitmap,
         'largeIconBitmapSource': AndroidBitmapSource.drawable.index,
       };
     } else if (largeIcon is FilePathAndroidBitmap) {
       return <String, Object>{
-        'largeIcon': largeIcon.bitmap,
+        'largeIcon': largeIcon!.bitmap,
         'largeIconBitmapSource': AndroidBitmapSource.filePath.index,
       };
     } else {
@@ -170,8 +170,8 @@ extension BigPictureStyleInformationMapper on BigPictureStyleInformation {
 }
 
 extension BigTexStyleInformationMapper on BigTextStyleInformation {
-  Map<String, Object> toMap() => _convertDefaultStyleInformationToMap(this)
-    ..addAll(<String, Object>{
+  Map<String, Object?> toMap() => _convertDefaultStyleInformationToMap(this)
+    ..addAll(<String, Object?>{
       'bigText': bigText,
       'htmlFormatBigText': htmlFormatBigText,
       'contentTitle': contentTitle,
@@ -182,37 +182,37 @@ extension BigTexStyleInformationMapper on BigTextStyleInformation {
 }
 
 extension InboxStyleInformationMapper on InboxStyleInformation {
-  Map<String, Object> toMap() => _convertDefaultStyleInformationToMap(this)
-    ..addAll(<String, Object>{
+  Map<String, Object?> toMap() => _convertDefaultStyleInformationToMap(this)
+    ..addAll(<String, Object?>{
       'contentTitle': contentTitle,
       'htmlFormatContentTitle': htmlFormatContentTitle,
       'summaryText': summaryText,
       'htmlFormatSummaryText': htmlFormatSummaryText,
-      'lines': lines ?? <String>[],
+      'lines': lines,
       'htmlFormatLines': htmlFormatLines
     });
 }
 
 extension MessagingStyleInformationMapper on MessagingStyleInformation {
-  Map<String, Object> toMap() => _convertDefaultStyleInformationToMap(this)
-    ..addAll(<String, Object>{
-      'person': person?.toMap(),
+  Map<String, Object?> toMap() => _convertDefaultStyleInformationToMap(this)
+    ..addAll(<String, Object?>{
+      'person': person.toMap(),
       'conversationTitle': conversationTitle,
       'groupConversation': groupConversation,
       'messages': messages
-          ?.map((m) => m?.toMap()) // ignore: always_specify_types
-          ?.toList()
+          ?.map((m) => m.toMap()) // ignore: always_specify_types
+          .toList()
     });
 }
 
 extension AndroidNotificationDetailsMapper on AndroidNotificationDetails {
-  Map<String, Object> toMap() => <String, Object>{
+  Map<String, Object?> toMap() => <String, Object?>{
         'icon': icon,
         'channelId': channelId,
         'channelName': channelName,
         'channelDescription': channelDescription,
         'channelShowBadge': channelShowBadge,
-        'channelAction': channelAction?.index,
+        'channelAction': channelAction.index,
         'importance': importance.value,
         'priority': priority.value,
         'playSound': playSound,
@@ -256,42 +256,42 @@ extension AndroidNotificationDetailsMapper on AndroidNotificationDetails {
         ..addAll(_convertNotificationSoundToMap(sound))
         ..addAll(_convertLargeIconToMap());
 
-  Map<String, Object> _convertStyleInformationToMap() {
+  Map<String, Object?> _convertStyleInformationToMap() {
     if (styleInformation is BigPictureStyleInformation) {
-      return <String, Object>{
+      return <String, Object?>{
         'style': AndroidNotificationStyle.bigPicture.index,
         'styleInformation':
-            (styleInformation as BigPictureStyleInformation)?.toMap(),
+            (styleInformation as BigPictureStyleInformation?)?.toMap(),
       };
     } else if (styleInformation is BigTextStyleInformation) {
-      return <String, Object>{
+      return <String, Object?>{
         'style': AndroidNotificationStyle.bigText.index,
         'styleInformation':
-            (styleInformation as BigTextStyleInformation)?.toMap(),
+            (styleInformation as BigTextStyleInformation?)?.toMap(),
       };
     } else if (styleInformation is InboxStyleInformation) {
-      return <String, Object>{
+      return <String, Object?>{
         'style': AndroidNotificationStyle.inbox.index,
         'styleInformation':
-            (styleInformation as InboxStyleInformation)?.toMap(),
+            (styleInformation as InboxStyleInformation?)?.toMap(),
       };
     } else if (styleInformation is MessagingStyleInformation) {
-      return <String, Object>{
+      return <String, Object?>{
         'style': AndroidNotificationStyle.messaging.index,
         'styleInformation':
-            (styleInformation as MessagingStyleInformation)?.toMap(),
+            (styleInformation as MessagingStyleInformation?)?.toMap(),
       };
     } else if (styleInformation is MediaStyleInformation) {
-      return <String, Object>{
+      return <String, Object?>{
         'style': AndroidNotificationStyle.media.index,
         'styleInformation':
-            (styleInformation as MediaStyleInformation)?.toMap(),
+            (styleInformation as MediaStyleInformation?)?.toMap(),
       };
     } else if (styleInformation is DefaultStyleInformation) {
-      return <String, Object>{
+      return <String, Object?>{
         'style': AndroidNotificationStyle.defaultStyle.index,
         'styleInformation':
-            (styleInformation as DefaultStyleInformation)?.toMap(),
+            (styleInformation as DefaultStyleInformation?)?.toMap(),
       };
     } else {
       return <String, Object>{
@@ -304,12 +304,12 @@ extension AndroidNotificationDetailsMapper on AndroidNotificationDetails {
   Map<String, Object> _convertLargeIconToMap() {
     if (largeIcon is DrawableResourceAndroidBitmap) {
       return <String, Object>{
-        'largeIcon': largeIcon.bitmap,
+        'largeIcon': largeIcon!.bitmap,
         'largeIconBitmapSource': AndroidBitmapSource.drawable.index,
       };
     } else if (largeIcon is FilePathAndroidBitmap) {
       return <String, Object>{
-        'largeIcon': largeIcon.bitmap,
+        'largeIcon': largeIcon!.bitmap,
         'largeIconBitmapSource': AndroidBitmapSource.filePath.index,
       };
     } else {
