@@ -22,7 +22,7 @@ extension MacOSNotificationAttachmentMapper on MacOSNotificationAttachment {
 }
 
 extension MacOSNotificationDetailsMapper on MacOSNotificationDetails {
-  Map<String, Object> toMap() => <String, Object>{
+  Map<String, Object?> toMap() => <String, Object?>{
         'presentAlert': presentAlert,
         'presentSound': presentSound,
         'presentBadge': presentBadge,
@@ -32,6 +32,6 @@ extension MacOSNotificationDetailsMapper on MacOSNotificationDetails {
         'threadIdentifier': threadIdentifier,
         'attachments': attachments
             ?.map((a) => a.toMap()) // ignore: always_specify_types
-            ?.toList()
+            .toList()
       };
 }
